+++ lib/isolator/adapters/background_jobs/resque_scheduler.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+Isolator.isolate :resque_scheduler,
+                 target: Resque.singleton_class,
+                 method_name: :enqueue_at,
+                 exception_class: Isolator::BackgroundJobError