+++ spec/spec_helper.rb
@@ -11,9 +11,7 @@ require "sidekiq/testing"
 Sidekiq::Testing.fake!
 
 require "resque"
-require "fakeredis"
-Resque.redis = Redis.new
-
+require "resque-scheduler"
 require "sucker_punch"
 
 require "net/http"